// src/enum/store-actions.enum.ts
export enum StoreActions {

    // burger store actions
    ADD_INGREDIENT = 'ADD INGREDIENT',
    REMOVE_INGREDIENT = 'REMOVE INGREDIENT',
    INITIALISE_BURGER = 'INITIALISE BURGER',

    // order store actions

    //order burger
    ORDER_BURGER = 'ORDER BURGER',
    BURGER_ORDER_INITIATED = 'BURGER ORDER INITIATED',
    BURGER_ORDER_COMPLETED = 'BURGER_ORDER_COMPLETED',

    // fetch orders
    ORDERS_LOADING = 'ORDERS LOADING',
    ORDERS_LOADED = 'ORDERS LOADED',
    GET_ORDERS = 'GET ORDERS',
    SAVE_ORDERS = 'SAVE ORDERS',

    // create draft order
    CREATE_DRAFT_ORDER = 'CREATE DRAFT ORDER',
    CLEAR_DRAFT_ORDER = 'CLEAR DRAFT ORDER'
}